Support for sending faster than the script is beeing called by cron

+++ setup/03_db_layout.sql
@@ -20,12 +20,12 @@ CREATE TABLE `load` (
 );
 
 CREATE TABLE `memory` (
- `mem_total` int(20) NOT NULL,
- `file_cache_size` int(20) NOT NULL,
- `used_programms` int(20) NOT NULL,
- `used_buffers` int(20) NOT NULL,
- `used_cache` int(20) NOT NULL,
- `free` int(20) NOT NULL,
+ `mem_total` bigint(20) NOT NULL,
+ `file_cache_size` bigint(20) NOT NULL,
+ `used_programms` bigint(20) NOT NULL,
+ `used_buffers` bigint(20) NOT NULL,
+ `used_cache` bigint(20) NOT NULL,
+ `free` bigint(20) NOT NULL,
 
  `server_id` int(10) unsigned NOT NULL,
  `time` bigint(20) NOT NULL,
@@ -37,10 +37,10 @@ CREATE TABLE `memory` (
 
 CREATE TABLE `network` (
  `interface` varchar(191) NOT NULL,
- `sent_bytes` int(20) NOT NULL,
- `received_bytes` int(20) NOT NULL,
- `sent_packets` int(20) NOT NULL,
- `received_packets` int(20) NOT NULL,
+ `sent_bytes` bigint(20) NOT NULL,
+ `received_bytes` bigint(20) NOT NULL,
+ `sent_packets` bigint(20) NOT NULL,
+ `received_packets` bigint(20) NOT NULL,
 
  `server_id` int(10) unsigned NOT NULL,
  `time` bigint(20) NOT NULL,
@@ -53,11 +53,11 @@ CREATE TABLE `network` (
 CREATE TABLE `diskusage` (
  `diskname` varchar(191) NOT NULL,
  `mounted_at` varchar(191) NOT NULL,
- `kbytes_all` int(20) NOT NULL,
- `kbytes_used` int(20) NOT NULL,
- `kbytes_reserved` int(20) NOT NULL,
- `inodes_all` int(20) NOT NULL,
- `inodes_used` int(20) NOT NULL,
+ `kbytes_all` bigint(20) NOT NULL,
+ `kbytes_used` bigint(20) NOT NULL,
+ `kbytes_reserved` bigint(20) NOT NULL,
+ `inodes_all` bigint(20) NOT NULL,
+ `inodes_used` bigint(20) NOT NULL,
 
  `server_id` int(10) unsigned NOT NULL,
  `time` bigint(20) NOT NULL,
@@ -69,10 +69,10 @@ CREATE TABLE `diskusage` (
 
 CREATE TABLE `diskio` (
  `diskname` varchar(191) NOT NULL,
- `read_io` int(20) NOT NULL,
- `read_sector` int(20) NOT NULL,
- `write_io` int(20) NOT NULL,
- `write_sector` int(20) NOT NULL,
+ `read_io` bigint(20) NOT NULL,
+ `read_sector` bigint(20) NOT NULL,
+ `write_io` bigint(20) NOT NULL,
+ `write_sector` bigint(20) NOT NULL,
 
  `server_id` int(10) unsigned NOT NULL,
  `time` bigint(20) NOT NULL,
@@ -84,12 +84,12 @@ CREATE TABLE `diskio` (
 
 CREATE TABLE `cpu` (
  `name` varchar(191) NOT NULL,
- `all` int(20) NOT NULL,
- `user` int(20) NOT NULL,
- `user_niced` int(20) NOT NULL,
- `kernel` int(20) NOT NULL,
- `io_wait` int(20) NOT NULL,
- `idle` int(20) NOT NULL,
+ `all` bigint(20) NOT NULL,
+ `user` bigint(20) NOT NULL,
+ `user_niced` bigint(20) NOT NULL,
+ `kernel` bigint(20) NOT NULL,
+ `io_wait` bigint(20) NOT NULL,
+ `idle` bigint(20) NOT NULL,
 
  `server_id` int(10) unsigned NOT NULL,
  `time` bigint(20) NOT NULL,
@@ -100,31 +100,31 @@ CREATE TABLE `cpu` (
 );
 
 CREATE TABLE `sql` (
- `bytes_received` int(20) NOT NULL,
- `bytes_sent` int(20) NOT NULL,
- `handler_commit` int(20) NOT NULL,
- `handler_delete` int(20) NOT NULL,
- `handler_update` int(20) NOT NULL,
- `handler_write` int(20) NOT NULL,
- `innodb_data_read` int(20) NOT NULL,
- `innodb_data_written` int(20) NOT NULL,
- `innodb_data_reads` int(20) NOT NULL,
- `innodb_data_writes` int(20) NOT NULL,
- `queries` int(20) NOT NULL,
- `connections` int(20) NOT NULL,
+ `bytes_received` bigint(20) NOT NULL,
+ `bytes_sent` bigint(20) NOT NULL,
+ `handler_commit` bigint(20) NOT NULL,
+ `handler_delete` bigint(20) NOT NULL,
+ `handler_update` bigint(20) NOT NULL,
+ `handler_write` bigint(20) NOT NULL,
+ `innodb_data_read` bigint(20) NOT NULL,
+ `innodb_data_written` bigint(20) NOT NULL,
+ `innodb_data_reads` bigint(20) NOT NULL,
+ `innodb_data_writes` bigint(20) NOT NULL,
+ `queries` bigint(20) NOT NULL,
+ `connections` bigint(20) NOT NULL,
 
- `innodb_buffer_pool_bytes_data` int(11) NOT NULL,
- `innodb_buffer_pool_pages_data` int(11) NOT NULL,
- `Innodb_buffer_pool_pages_dirty` int(11) NOT NULL,
- `Innodb_buffer_pool_bytes_dirty` int(11) NOT NULL,
- `innodb_buffer_pool_pages_free` int(11) NOT NULL,
- `innodb_buffer_pool_pages_flushed` int(11) NOT NULL,
- `innodb_mem_dictionary` int(11) NOT NULL,
+ `innodb_buffer_pool_bytes_data` bigint(20) NOT NULL,
+ `innodb_buffer_pool_pages_data` bigint(20) NOT NULL,
+ `Innodb_buffer_pool_pages_dirty` bigint(20) NOT NULL,
+ `Innodb_buffer_pool_bytes_dirty` bigint(20) NOT NULL,
+ `innodb_buffer_pool_pages_free` bigint(20) NOT NULL,
+ `innodb_buffer_pool_pages_flushed` bigint(20) NOT NULL,
+ `innodb_mem_dictionary` bigint(20) NOT NULL,
 
  `qcache_free_memory` int(11) NOT NULL,
- `qcache_hits` int(20) NOT NULL,
- `qcache_inserts` int(20) NOT NULL,
- `qcache_not_cached` int(20) NOT NULL,
+ `qcache_hits` bigint(20) NOT NULL,
+ `qcache_inserts` bigint(20) NOT NULL,
+ `qcache_not_cached` bigint(20) NOT NULL,
  `qcache_total_blocks` int(11) NOT NULL,
 
  `server_id` int(10) unsigned NOT NULL,
